Refactor implement interface to push code into service
